+++ src/consumer/messages/service/messages.py
@@ -1,13 +1,20 @@
 # SPDX-License-Identifier: AGPL-3.0-only
 # SPDX-FileCopyrightText: 2024 Univention GmbH
-
+import asyncio
 import logging
 from typing import List, Optional
 
 from consumer.port import ConsumerPort
 
 from consumer.subscriptions.service.subscription import SubscriptionService
-from shared.models import FillQueueStatus, MQMessage, Message
+from shared.models import (
+    MessageProcessingStatusReport,
+    MessageProcessingStatus,
+    PublisherName,
+    ProvisioningMessage,
+    FillQueueStatus,
+    Message,
+)
 
 
 PREFILL_SUBJECT_TEMPLATE = "prefill_{subject}"
@@ -37,7 +44,7 @@ class MessageService:
         pop: bool,
         timeout: float = 5,
         skip_prefill: Optional[bool] = False,
-    ) -> Optional[MQMessage]:
+    ) -> Optional[ProvisioningMessage]:
         """Retrieve the first message from the subscription's stream.
 
         :param str subscription_name: Name of the subscription.
@@ -58,7 +65,7 @@ class MessageService:
         count: int,
         pop: bool,
         skip_prefill: Optional[bool],
-    ) -> List[MQMessage]:
+    ) -> List[ProvisioningMessage]:
         """Return messages from a given queue.
 
         :param str subscription_name: Name of the subscription.
@@ -95,7 +102,7 @@ class MessageService:
 
     async def get_messages_from_main_queue(
         self, subject: str, timeout: float, count: int, pop: bool
-    ) -> List[MQMessage]:
+    ) -> List[ProvisioningMessage]:
         self.logger.info(
             "Getting the messages for the '%s' from the main queue", subject
         )
@@ -103,7 +110,7 @@ class MessageService:
 
     async def get_messages_from_prefill_queue(
         self, subject: str, timeout: float, count: int, pop: bool
-    ) -> List[MQMessage]:
+    ) -> List[ProvisioningMessage]:
         self.logger.info(
             "Getting the messages for the '%s' from the prefill queue", subject
         )
@@ -120,13 +127,28 @@ class MessageService:
                 await self._port.delete_stream(prefill_subject)
         return messages
 
-    async def remove_message(self, msg: MQMessage):
-        """Remove a message from the subscription's queue.
+    async def post_messages_status(
+        self, subscription_name: str, reports: List[MessageProcessingStatusReport]
+    ):
+        tasks = [
+            self.delete_message(subscription_name, report)
+            for report in reports
+            if report.status == MessageProcessingStatus.ok
+        ]
+        # Gather all tasks and run them concurrently
+        await asyncio.gather(*tasks)
 
-        :param msg: fetched message.
-        """
+    async def delete_message(
+        self, subscription_name: str, report: MessageProcessingStatusReport
+    ):
+        """Delete the messages from the subscriber's queue."""
 
-        await self._port.remove_message(msg)
+        if report.publisher_name == PublisherName.udm_pre_fill:
+            stream_name = PREFILL_SUBJECT_TEMPLATE.format(subject=subscription_name)
+        else:
+            stream_name = subscription_name
+
+        await self._port.delete_message(stream_name, report.message_seq_num)
 
     async def create_prefill_stream(self, subscription_name: str):
         # delete the previously created stream if it exists