+++ tests/conftest.py
@@ -17,18 +17,31 @@ from consumer.port import ConsumerPort
 from consumer.main import app
 from events.port import EventsPort
 from shared.adapters.nats_adapter import NatsKVAdapter, NatsMQAdapter
-from shared.models import Message
-from shared.models.queue import MQMessage
-from shared.models.queue import PrefillMessage
+from shared.models import (
+    Message,
+    MessageProcessingStatusReport,
+    MessageProcessingStatus,
+    MQMessage,
+    PublisherName,
+    ProvisioningMessage,
+    PrefillMessage,
+)
 
 REALM = "udm"
 TOPIC = "groups/group"
 BODY = {"new": {"New": "Object"}, "old": {"Old": "Object"}}
-PUBLISHER_NAME = "udm-listener"
+PUBLISHER_NAME = PublisherName.udm_listener
 REALM_TOPIC = [REALM, TOPIC]
-REALMS_TOPICS = [[REALM, TOPIC]]
+REALMS_TOPICS = [(REALM, TOPIC)]
 REALMS_TOPICS_STR = f"{REALM}:{TOPIC}"
 SUBSCRIPTION_NAME = "0f084f8c-1093-4024-b215-55fe8631ddf6"
+REPLY = f"$JS.ACK.stream:{SUBSCRIPTION_NAME}.durable_name:{SUBSCRIPTION_NAME}.1.1.1.1699615014739091916.0"
+
+REPORT = MessageProcessingStatusReport(
+    status=MessageProcessingStatus.ok,
+    message_seq_num=1,
+    publisher_name=PublisherName.udm_listener,
+)
 
 SUBSCRIPTION_INFO = {
     "name": SUBSCRIPTION_NAME,
@@ -49,6 +62,15 @@ PREFILL_MESSAGE = PrefillMessage(
     realms_topics=REALMS_TOPICS,
     subscription_name=SUBSCRIPTION_NAME,
 )
+PROVISIONING_MESSAGE = ProvisioningMessage(
+    publisher_name=PUBLISHER_NAME,
+    ts=datetime(2023, 11, 9, 11, 15, 52, 616061),
+    realm=REALM,
+    topic=TOPIC,
+    body=BODY,
+    sequence_number=1,
+    num_delivered=1,
+)
 
 FLAT_BASE_MESSAGE = {
     "publisher_name": PUBLISHER_NAME,
@@ -65,6 +87,7 @@ FLAT_PREFILL_MESSAGE["realms_topics"] = REALMS_TOPICS
 
 MSG = Msg(
     _client="nats",
+    reply=REPLY,
     data=json.dumps(FLAT_MESSAGE).encode(),
     _metadata=Msg.Metadata(
         sequence=Msg.Metadata.SequencePair(consumer=5, stream=5),
@@ -78,6 +101,7 @@ MSG = Msg(
 )
 MSG_PREFILL = Msg(
     _client="nats",
+    reply=REPLY,
     data=json.dumps(FLAT_PREFILL_MESSAGE).encode(),
     _metadata=Msg.Metadata(
         sequence=Msg.Metadata.SequencePair(consumer=5, stream=5),
@@ -91,6 +115,7 @@ MSG_PREFILL = Msg(
 )
 MSG_PREFILL_REDELIVERED = Msg(
     _client="nats",
+    reply=REPLY,
     data=json.dumps(FLAT_PREFILL_MESSAGE).encode(),
     _metadata=Msg.Metadata(
         sequence=Msg.Metadata.SequencePair(consumer=5, stream=5),
@@ -104,7 +129,12 @@ MSG_PREFILL_REDELIVERED = Msg(
 )
 
 MQMESSAGE = MQMessage(
-    subject="", reply="", data=FLAT_MESSAGE, headers=None, num_delivered=1
+    subject="",
+    reply=REPLY,
+    data=FLAT_MESSAGE,
+    headers=None,
+    num_delivered=1,
+    sequence_number=1,
 )
 
 MQMESSAGE_PREFILL = deepcopy(MQMESSAGE)